# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Polaris Developer API
  version: 0.1.0-beta.1
  description: >-
    Public developer contract for the Synara control plane. Operations marked
    route-only participate in exposure conformance but are not yet eligible for SDK generation.
servers:
  - url: /
paths:
  /v1/artifacts/{artifactID}:
    delete:
      operationId: deleteArtifact
      summary: Idempotently delete an Artifact and revoke its grants
      x-synara-exposure: public-beta
      x-synara-contract-status: codegen-ready
      x-synara-request-body: none
      parameters:
        - name: artifactID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - $ref: "#/components/parameters/IdempotencyKey"
      security:
        - ServiceAccountBearer: []
      responses:
        "204":
          description: Artifact deleted, or already deleted by the same authorized caller
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      operationId: getArtifact
      summary: Get Artifact metadata
      x-synara-exposure: public-beta
      x-synara-contract-status: codegen-ready
      parameters:
        - name: artifactID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      security:
        - ServiceAccountBearer: []
      responses:
        "200":
          description: Artifact metadata without payload bytes or grant credentials
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/artifacts/{artifactID}/complete:
    post:
      operationId: completeArtifact
      summary: Confirm an uploaded Artifact by size, digest, and media type
      x-synara-exposure: public-beta
      x-synara-contract-status: codegen-ready
      parameters:
        - name: artifactID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompleteArtifactRequest" }
      security:
        - ServiceAccountBearer: []
      responses:
        "200":
          description: Ready Artifact; an identical confirmation is intrinsically replay-safe
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/artifacts/{artifactID}/download:
    post:
      operationId: downloadArtifact
      summary: Issue a short-lived Artifact download grant
      x-synara-exposure: public-beta
      x-synara-contract-status: codegen-ready
      x-synara-request-body: none
      parameters:
        - name: artifactID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      security:
        - ServiceAccountBearer: []
      responses:
        "200":
          description: Short-lived download grant; payload bytes remain outside the JSON API
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactDownloadGrant" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/sessions/{sessionID}/artifacts:
    get:
      operationId: listArtifacts
      summary: List Artifact metadata for a Session
      x-synara-exposure: public-beta
      x-synara-contract-status: codegen-ready
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          schema: { type: string, minLength: 1 }
      security:
        - ServiceAccountBearer: []
      responses:
        "200":
          description: Bounded Artifact metadata page ordered newest first
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactPage" }
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createArtifact
      summary: Create pending Artifact metadata and a short-lived upload grant
      x-synara-exposure: public-beta
      x-synara-contract-status: codegen-ready
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateArtifactRequest" }
      security:
        - ServiceAccountBearer: []
      responses:
        "201":
          description: Pending Artifact with a fresh short-lived grant, or ready metadata on replay
          headers:
            Idempotency-Replayed: { $ref: "#/components/headers/IdempotencyReplayed" }
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactUploadGrant" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    ServiceAccountBearer:
      type: http
      scheme: bearer
      bearerFormat: syna_sa_...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated opaque key used to replay this mutation safely.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  headers:
    IdempotencyReplayed:
      description: True when the stored response for the Idempotency-Key was replayed.
      schema:
        type: boolean
    RateLimitLimit:
      description: Per-key request limit for the current UTC minute window.
      schema:
        type: integer
        minimum: 1
    RateLimitRemaining:
      description: Remaining admitted requests for the key in the current window.
      schema:
        type: integer
        minimum: 0
    RateLimitReset:
      description: Whole seconds until the current UTC minute window resets.
      schema:
        type: integer
        minimum: 1
  responses:
    ErrorResponse:
      description: Stable Polaris API error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    UUID:
      type: string
      format: uuid
...
    Artifact:
      type: object
      additionalProperties: false
      required:
        - id
        - tenantId
        - organizationId
        - projectId
        - sessionId
        - executionId
        - kind
        - status
        - originalName
        - contentType
        - sizeBytes
        - sha256
        - createdByType
        - createdById
        - readyAt
        - createdAt
        - expiresAt
        - deletedAt
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        tenantId: { $ref: "#/components/schemas/UUID" }
        organizationId: { $ref: "#/components/schemas/UUID" }
        projectId: { $ref: "#/components/schemas/UUID" }
        sessionId: { $ref: "#/components/schemas/UUID" }
        executionId:
          oneOf: [{ $ref: "#/components/schemas/UUID" }, { type: "null" }]
        kind:
          type: string
          enum: [attachment, generated_file, terminal_log, diff, workspace_snapshot, checkpoint, memory]
        status: { type: string, enum: [pending, ready, deleting, deleted, failed] }
        originalName: { type: [string, "null"] }
        contentType: { type: [string, "null"] }
        sizeBytes: { type: [integer, "null"], format: int64, minimum: 0 }
        sha256:
          oneOf: [{ type: string, pattern: "^[0-9a-f]{64}$" }, { type: "null" }]
        createdByType: { type: string, enum: [user, service_account, worker] }
        createdById: { $ref: "#/components/schemas/UUID" }
        readyAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
        deletedAt: { type: [string, "null"], format: date-time }
    CreateArtifactRequest:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind:
          type: string
          enum: [attachment, generated_file, terminal_log, diff, workspace_snapshot]
        originalName: { type: [string, "null"], maxLength: 500 }
        executionId:
          oneOf: [{ $ref: "#/components/schemas/UUID" }, { type: "null" }]
        expiresAt: { type: [string, "null"], format: date-time }
    CompleteArtifactRequest:
      type: object
      additionalProperties: false
      required: [sizeBytes, sha256, contentType]
      properties:
        sizeBytes: { type: integer, format: int64, minimum: 0 }
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        contentType: { type: string, minLength: 1, maxLength: 255 }
    ArtifactUploadGrant:
      type: object
      additionalProperties: false
      required: [artifact, uploadRequired]
      properties:
        artifact: { $ref: "#/components/schemas/Artifact" }
        uploadRequired: { type: boolean }
        method: { type: string, enum: [PUT] }
        url: { type: string, minLength: 1 }
        headers: { type: object, additionalProperties: { type: string } }
        expiresAt: { type: string, format: date-time }
    ArtifactPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Artifact" }
        nextCursor: { type: [string, "null"] }
    ArtifactDownloadGrant:
      type: object
      additionalProperties: false
      required: [artifact, url, expiresAt]
      properties:
        artifact: { $ref: "#/components/schemas/Artifact" }
        url:
          type: string
          minLength: 1
          description: Short-lived relative local-content URL or absolute presigned object-store URL.
        expiresAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details:
              oneOf:
                - type: object
                  additionalProperties: true
                - type: "null"
...